consolodates part 1 tests into regression suite
I know that what I have in part 1 isn't going to work going forward,
but I also know that I don't want to break it. Since the content of
the tests was good but the code itself was bare-bones and unhelpful,
I'm opting to consolodate these into a single chunk of regression
tests.