reworks Operation to take full list of instructions
We have to be able to look at the entire list of instructions if we're
going to be able to pull out all the values we need for our
operations.

Note that I'm not thrilled with the tests I have right now. I think it
might abstract a bit too much, as now the only context you have in a
given test operation block is about the index of the item you are
looking at, where the actual instruction list is pretty far away,
LOC-speaking. I might come back and change this around, but given what
I know now I think it's workable as is so I won't put any more time
into it just yet.