starts moving to domain language
In part 2 of the question I learned that the instructions that get fed
into the program are stored in memory, and the memory is the thing
that is being manipulated as we run the program (and it's the thing we
update and report on). This is still a little uneasy, but using the
domain language is going to help in the long run, I believe.