adds class method to find noun/verb
Two things are going on here. The most obvious is that I am adding
some functionality to answer the second part of the actual
problem/question. But perhaps a more important part is that we are
starting to refine the domain language a bit more. Now we have names
for the second and third places in the program's memory, and we can
refer to those by name in operations, thereby moving some domain
knowledge _out of_ callers and driving it into the objects themselves.